Initialize strtab base and base_cfg correctly. Still not able to enable SMMU in bypass mode.

+++ sys/dev/iommu/smmu_var.h
@@ -66,7 +66,7 @@ struct smmu_cmdq_entry {
 struct smmu_strtab {
 	void *addr;
 	uint64_t base;
-	uint64_t base_cfg;
+	uint32_t base_cfg;
 };
 
 struct smmu_softc {
